requirements.txt
aiogram==2.19
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
environs==9.5.0
frozenlist==1.3.0
idna==3.3
marshmallow==3.14.1
multidict==6.0.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
soupsieve==2.3.1
typing-extensions==4.1.1
urllib3==1.26.8
yarl==1.7.2
psycopg2-binary==2.9.3